Remove foreign key relation from BSN
The current foreign key relation from Subscriptions to BSNMutation
caused issues in the process to fill the BSNMutation table. This has
been changed to match the NewResidents table.

Mutation_date was renamed to inserted_at, which was already applied in
the test database. Since we're not deployed yet, I've changed the
migration.